refactor: board 타입 정리
